Current Work/Enrichment.Inference.Engine — File Pack v2/pyproject.toml
[tool.poetry]
name = "l9-enrich-inference-engine"
version = "0.3.0"
description = "L9 Cognitive Enrichment Engine — Layer 2 ENRICH service"
authors = ["L9 Labs <eng@l9labs.ai>"]
python = "^3.12"

[tool.poetry.dependencies]
python          = "^3.12"
fastapi         = "^0.115"
uvicorn         = {extras = ["standard"], version = "^0.32"}
pydantic        = "^2.9"
pydantic-settings = "^2.6"
httpx           = "^0.28"
structlog       = "^24.4"
pyyaml          = "^6.0"
sqlalchemy      = {extras = ["asyncio"], version = "^2.0"}
asyncpg         = "^0.30"
aioredis        = "^2.0"
alembic         = "^1.14"
jinja2          = "^3.1"
python-multipart = "^0.0.12"

[tool.poetry.group.dev.dependencies]
pytest          = "^8.3"
pytest-asyncio  = "^0.24"
pytest-cov      = "^6.0"
httpx           = "^0.28"
ruff            = "^0.8"
mypy            = "^1.13"
testcontainers  = {extras = ["neo4j"], version = "^4.8"}
locust          = "^2.32"

[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode   = "auto"
testpaths      = ["tests"]
addopts        = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select  = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore  = ["E501", "B008"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S307"]

[tool.mypy]
python_version = "3.12"
strict         = false
ignore_missing_imports = true
